refactor(vim): rename packer_init -> core.packer
