+++ editor/neovim/init.lua
@@ -1,5 +1,5 @@
-require('packer_init')
-require('core/options')
-require('core/keymaps')
-require('core/autocommands')
-require('legacy/navigation')
+require('core.packer')
+require('core.options')
+require('core.keymaps')
+require('core.autocommands')
+require('legacy.navigation')
